refactor(devices): split HomeViewModel app-chip machinery into a partial file
Move the contiguous app-indicator chip block (per-card reconcile, chip ordering, hidden-apps
lookup, hide/close/terminate) out of HomeViewModel.cs into HomeViewModel.Apps.cs (2516 -> 1977 +
556). Plain methods only; the source-generated [ObservableProperty]/[RelayCommand] members stay in
the main partial (XAML markup compiler constraint).
